Add ofxPanel mouse detection Inside (). Repair ofxSlider text display problems.

+++ addons/ofxGui/src/ofxSlider.cpp
@@ -142,7 +142,7 @@ void ofxSlider<Type>::render(){
 
 	bindFontTexture();
 #ifdef SUPPORT_FONTSTASH
-	string valStr = ofToString((int)value);
+	string valStr = ofToString((float)value);
 	unicodeFont.draw(getName(),fontSize, b.x + textPadding, b.y + b.height / 2 + fontSize/2 - (2 * (fontSize/12)));
 	unicodeFont.draw(valStr,fontSize, b.x + b.width - textPadding - getTextBoundingBox(valStr,0,0).width, b.y + b.height / 2 + fontSize/2 - (2 * (fontSize/12)));
 #else
